chore: Move ConversationVerifiedSystemMessageSectionDescription into its own file
